# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Fsirst API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/teams/": {
            "post": {
                "tags": [
                    "팀"
                ],
                "description": "팀 생성",
                "operationId": "fe0ee24d5e23fe8d1d33e40ad10107e4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "팀이름",
                                        "type": "string",
                                        "example": "갓챠"
                                    },
                                    "contents": {
                                        "description": "팀소개",
                                        "type": "string",
                                        "example": "내용"
                                    },
                                    "region": {
                                        "description": "지역",
                                        "type": "string",
                                        "example": "강남구"
                                    },
                                    "limit_person": {
                                        "description": "정원",
                                        "type": "string",
                                        "example": "1234"
                                    },
                                    "sex": {
                                        "description": "성별",
                                        "type": "string",
                                        "example": "혼성:0, 남자:1, 여자:2"
                                    },
                                    "min_age": {
                                        "description": "최소나이",
                                        "type": "string",
                                        "example": "2000"
                                    },
                                    "max_age": {
                                        "description": "최대나이",
                                        "type": "string",
                                        "example": "1993"
                                    },
                                    "files[]": {
                                        "description": "이미지 input name=files[]",
                                        "type": "file",
                                        "example": "file.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
        "/api/teams/update/{tid}": {
            "post": {
                "tags": [
                    "팀"
                ],
                "description": "팀 수정",
                "operationId": "59944b863dda4aff9b3aa7484936be2d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "팀이름",
                                        "type": "string",
                                        "example": "갓챠"
                                    },
                                    "contents": {
                                        "description": "팀소개",
                                        "type": "string",
                                        "example": "내용"
                                    },
                                    "region": {
                                        "description": "지역",
                                        "type": "string",
                                        "example": "강남구"
                                    },
                                    "limit_person": {
                                        "description": "정원",
                                        "type": "string",
                                        "example": "1234"
                                    },
                                    "sex": {
                                        "description": "성별",
                                        "type": "string",
                                        "example": "혼성:0, 남자:1, 여자:2"
                                    },
                                    "min_age": {
                                        "description": "최소나이",
                                        "type": "string",
                                        "example": "2000"
                                    },
                                    "max_age": {
                                        "description": "최대나이",
                                        "type": "string",
                                        "example": "1990"
                                    },
                                    "confirm_m": {
                                        "description": "운영진 승인(Y/N)",
                                        "type": "string",
                                        "example": "Y"
                                    },
                                    "files[]": {
                                        "description": "이미지 input name=files[]",
                                        "type": "file",
                                        "example": "file.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
        "/api/teams": {
            "get": {
                "tags": [
                    "팀"
                ],
                "description": "팀 리스트 불러오기 [per_page=>페이지 당 레코드 갯수] [page =>보여줄 페이지]",
                "operationId": "96d86695ce86b7ef00a0417e16d08e2c",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
        "/api/teams/searchTeams": {
            "post": {
                "tags": [
                    "팀"
                ],
                "description": "팀 검색",
                "operationId": "7ac2d9a3dd4810d338bdcdd688ca11ce",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "팀이름",
                                        "type": "string",
                                        "example": "갓챠"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
        "/api/teams/{$tid}": {
            "get": {
                "tags": [
                    "팀"
                ],
                "description": "팀 상세 불러오기",
                "operationId": "2ad53690620ecde4acaff1502420424e",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
        "/api/teams/detail-match/{$tid}": {
            "get": {
                "tags": [
                    "팀"
                ],
                "description": "팀 경기 상세 불러오기",
                "operationId": "ee00833541e10ec8137d522367ff238a",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
        "/api/teams/delete-team/{tid}": {
            "post": {
                "tags": [
                    "팀"
                ],
                "description": "팀 삭제",
                "operationId": "f17aabfa46bcd7eddd7021f8199921fa",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        }
    }
}